major refactor of playback-coordinator.service music-player.service and play-list-logic.service to consolidate code that was stringing across all three and to help reduce duplication code or bloat
